### 2020-12-11 #2899 @whedon
[PRE REVIEW]: Riroriro: Simulating gravitational waves and calculating SNRs in Python
**Submitting author:** @wvanzeist (<a href="http://orcid.org/0000-0002-4231-656X">Wouter van Zeist</a>)
**Repository:** <a href="https://github.com/wvanzeist/riroriro" target ="_blank">https://github.com/wvanzeist/riroriro</a>
**Version:** v0.1.4
