### 2020-12-11 #2899 @Kevin-Mattheus-Moerman
@wvanzeist since the core functionality appears to be under 1000 lines of code I have flagged this submission for an additional scope check by the editorial board. In particular our board will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. This is standard for submissions of this size. We should get back to you within about a week.
