### 2020-12-17 #2899 @kthyng
Hi @wvanzeist, thanks for your submission. In discussion with the editorial board at JOSS, a question came up. Your package apparently implements a subset of what is in this package (bilby: https://github.com/lscsoft/bilby). Could you tell us what distinguishes your package from bilby? This is especially important since Riroriro is pretty small for lines of code.
