### 2020-12-18 #2899 @kthyng
I'm keeping the scope query in place while we hear more about this submission, but I want to bring in @dfm who would be the likely handling editor.
