### 2020-12-19 #2899 @dfm
@wvanzeist In case you missed it, please respond to @kthyng's [question above](https://github.com/openjournals/joss-reviews/issues/2899#issuecomment-747693470) when you get a chance. Thanks!
