### 2020-12-20 #2899 @wvanzeist
> Hi @wvanzeist, thanks for your submission. In discussion with the editorial board at JOSS, a question came up. Your package apparently implements a subset of what is in this package (bilby: https://github.com/lscsoft/bilby). Could you tell us what distinguishes your package from bilby? This is especially important since Riroriro is pretty small for lines of code.

Thank you for your comment. Work on Riroriro actually began before Bilby was published, which is why we hadn't come across it in our initial literature review of gravitational wave models. Since it is a related package and notably open-source and in Python, I will add a mention about it in the Riroriro paper.

However, Riroriro's functionality is not the same as that of Bilby. Bilby and Riroriro approach the topic of gravitational waves from distinct, almost opposite ways: Bilby is focused on analysing and performing parameter estimation via Bayesian inference on gravitational wave strain data that the user has to input from detectors or other software like LALSimulation; while Riroriro models, simulates and outputs gravitational wave strain data itself (forward modelling) and can analyse this data in terms of detectability, and can also be used as an educational tool for students learning about gravitational waves.

Does that clarify the purpose of Riroriro?
