### 2020-12-21 #2899 @dfm
@wvanzeist: Thanks for the info. This sounds good to me! Here at JOSS we're less interested in the educational angle, but I think that the forward modeling aspect is a good fit. Can you ping me once you've added a section clarifying this to the paper and then I'll start recruiting reviewers? This might be a little slow given the holidays but I'll keep you posted.
