### 2020-12-22 0b12af2
Mentioning GW parameter estimation packages in paper
And how they are distinct from Riroriro.
