+++ paper/paper.md
@@ -30,7 +30,9 @@ The gravitational waveform simulation of `Riroriro` is based upon the methods of
 
 Gravitational waves have long been an area of research in astronomy, and particularly since the first observation of gravitational waves was announced in 2016 [@gw150914discovery] this area has seen a lot of research activity. Observations of gravitational waves from binary mergers can provide unique information about their progenitors and stellar populations, especially when combined with electromagnetic observations in the field called multi-messenger astronomy. A major factor in the successful detection and analysis of gravitational wave signals is the creation of simulations of such signals which observed data can be compared to. Because of this, multiple gravitational wave models have been created over the years.
 
-The gravitational wave observatories LIGO/Virgo use their own models as templates, and various other research groups have also created models. For example, prior to @huerta2017 and @buskirk2019 there were other simulations of gravitational waves using similar methods of post-Newtonian expansions and/or limited numerical relativity simulations [@apostolatos1995; @buonanno2003; @ajith2007], though these had less mathematical accuracy than the later models. There are also models using alternative methods to the common post-Newtonian expansions, such as representing the two-body system of merging binaries as an effective one-body system [@buonanno1999], solving full numerical relativity equations but only at limits of the binary interaction and extrapolating for the rest of the waveform [@baker2002a; @baker2002b] or using numerically greedy algorithms to predict the shape of a wide range of waveforms based on a small number of given waveforms [@field2014]. However, for the most part these articles do not go into much detail about the computational implementation of their simulations and have not published their full database or its source code. Some have published it but in a language that is not free to use, like the original model of @buskirk2019 which is written in `Mathematica`.
+The gravitational wave observatories LIGO/Virgo use their own models as templates, and various other research groups have also created models. For example, prior to @huerta2017 and @buskirk2019 there were other simulations of gravitational waves using similar methods of post-Newtonian expansions and/or limited numerical relativity simulations [@apostolatos1995; @buonanno2003; @ajith2007], though these had less mathematical accuracy than the later models. There are also models using alternative methods to the common post-Newtonian expansions, such as representing the two-body system of merging binaries as an effective one-body system [@buonanno1999], solving full numerical relativity equations but only at limits of the binary interaction and extrapolating for the rest of the waveform [@baker2002a; @baker2002b] or using numerically greedy algorithms to predict the shape of a wide range of waveforms based on a small number of given waveforms [@field2014].
+
+However, for the most part these articles do not go into much detail about the computational implementation of their simulations and have not published their full database or its source code. Some have published it but in a language that is not free to use, like the original model of @buskirk2019 which is written in Mathematica. There are some open-source software packages in free-to-use programming languages, like `Bilby` [@bilby] and `PyCBC` [@pycbc] in Python, but these are more focused on performing analysis and parameter estimation on given signals rather than the forward modelling of signals performed by `Riroriro` and the models mentioned in the previous paragraph.
 
 `Riroriro` combines areas covered by previous models (such as gravitational wave simulation, SNR calculation, horizon distance calculation) into a single package with broader scope and versatility in Python, a programming language that is ubiquitous in astronomy. Aside from being a research tool, `Riroriro` is also designed to be easy to use and modify, and it can also be used as an educational tool for students learning about gravitational waves.
 
