+++ paper/paper.bib
@@ -219,7 +219,7 @@ archivePrefix = {arXiv},
         }, L.~A.~S. and {Taylor}, G. and {Ng}, M. and {Greis}, S.~M.~L. and
          {Bray}, J.~C.},
         title = "{Binary Population and Spectral Synthesis Version 2.1: Construction, observational verification, and new results}",
-      journal = {Publ. Astron. Soc. Australia},
+      journal = {Publ. Astron. Soc. Aust.},
      keywords = {binaries: general, galaxies: evolution, galaxies: stellar content, methods: numerical, stars: evolution, stars: statistics, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2017,
         month = nov,
@@ -294,4 +294,41 @@ archivePrefix = {arXiv},
 	author = {{Ghodla}, Sohan and {van Zeist}, Wouter G.~J. and {Eldridge}, J.~J.},
 	title = {Predicting the O3 gravitational wave mass distribution with BPASS},
 	year = "forthcoming"
+}
+
+@ARTICLE{bilby,
+       author = {{Ashton}, Gregory and {H{\"u}bner}, Moritz and {Lasky}, Paul D. and {Talbot}, Colm and {Ackley}, Kendall and {Biscoveanu}, Sylvia and {Chu}, Qi and {Divakarla}, Atul and {Easter}, Paul J. and {Goncharov}, Boris and {Hernandez Vivanco}, Francisco and {Harms}, Jan and {Lower}, Marcus E. and {Meadors}, Grant D. and {Melchor}, Denyz and {Payne}, Ethan and {Pitkin}, Matthew D. and {Powell}, Jade and {Sarin}, Nikhil and {Smith}, Rory J.~E. and {Thrane}, Eric},
+        title = "{BILBY: A user-friendly Bayesian inference library for gravitational-wave astronomy}",
+      journal = {ApJS},
+     keywords = {gravitational waves, methods: data analysis, methods: statistical, stars: black holes, stars: neutron, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
+         year = 2019,
+        month = apr,
+       volume = {241},
+       number = {2},
+          eid = {27},
+        pages = {27},
+          doi = {10.3847/1538-4365/ab06fc},
+archivePrefix = {arXiv},
+       eprint = {1811.02042},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ApJS..241...27A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{pycbc,
+       author = {{Biwer}, C.~M. and {Capano}, Collin D. and {De}, Soumi and {Cabero}, Miriam and {Brown}, Duncan A. and {Nitz}, Alexander H. and {Raymond}, V.},
+        title = "{PyCBC Inference: A Python-based parameter estimation toolkit for compact binary coalescence signals}”,
+      journal = {Publ. Astron. Soc. Pac.},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies, General Relativity and Quantum Cosmology},
+         year = 2019,
+        month = feb,
+       volume = {131},
+       number = {996},
+        pages = {024503},
+          doi = {10.1088/1538-3873/aaef0b},
+archivePrefix = {arXiv},
+       eprint = {1807.10312},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019PASP..131b4503B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
