+++ paper/paper.bib
@@ -317,7 +317,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{pycbc,
        author = {{Biwer}, C.~M. and {Capano}, Collin D. and {De}, Soumi and {Cabero}, Miriam and {Brown}, Duncan A. and {Nitz}, Alexander H. and {Raymond}, V.},
-        title = "{PyCBC Inference: A Python-based parameter estimation toolkit for compact binary coalescence signals}”,
+        title = "{PyCBC Inference: A Python-based parameter estimation toolkit for compact binary coalescence signals}",
       journal = {Publ. Astron. Soc. Pac.},
      keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies, General Relativity and Quantum Cosmology},
          year = 2019,
