### 2020-12-22 b84b73a
Clarifying sentence in paper
