### 2020-12-24 #2899 @wvanzeist
Looking at the list of reviewers, the following people have both "gravitational waves" and "Python" as keywords and thus might be suitable: duncanmmacleod, johnveitch, cmbiwer.
