### 2020-12-30 5fa13f3
Adding success messages for test functions
Previously, they just completed silently - with messages it is clearer for the user.
