+++ riroriro/tests/tests_mergersecondfuns.py
@@ -14,3 +14,5 @@ def test_mergersecondfuns_errors():
     
     with pytest.raises(AssertionError):
         me2.merger_phase_calculation(1.4,1,[0,2],[0.2])
+        
+    print('test_mergersecondfuns_errors completed successfully')
