+++ paper/paper.md
@@ -16,7 +16,7 @@ authors:
 affiliations:
   - name: Department of Physics, University of Auckland, New Zealand
     index: 1
-date: 2020
+date: 2021
 bibliography: paper.bib
 ---
 
