+++ licenses/LICENSE.rst
@@ -1,4 +1,4 @@
-Copyright 2020 Wouter van Zeist
+Copyright 2020-2021 Wouter van Zeist
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
