+++ riroriro/tests/tests_snrcalculatorfuns.py
@@ -31,7 +31,7 @@ def test_snrcalculatorfuns_numerical(inputarray,findchirp_array,noisearray):
     For inputarray and findchirp_array, please use the files
     'example_1.4_0.8.dat' and
     'findchirp_traces/mergerwave_1.40000_0.800000.dat', respectively, which are
-    provided in the riroriro_tutorials directory.
+    provided in the riroriro_tutorials repository.
     For noisearray, please use the file 'o3_l1.txt' from
     https://dcc.ligo.org/LIGO-T1500293/public
     """
