+++ docs/index.rst
@@ -16,3 +16,4 @@ This is the documentation for the functions of riroriro. Tutorials for the usage
   riroriro/snrcalculatorfuns.rst
   riroriro/horizondistfuns.rst
   riroriro/detectabilityfuns.rst
+  riroriro/tests.rst
