### 2021-01-12 9c2cc62
Redirecting function for backwards compatibility
A waveform_arrayer_4col() and redirecting function waveform_arrayer() should also be added at some point.
