### 2021-01-13 c5bf450
Fixing issue where polarisation amplitudes were not continuous at switching point
