### 2021-01-15 fed700a
Correcting constants in polarisation matching
