### 2021-01-19 7b1f7e3
Polarisation version of waveform_arrayer
And appropriate redirecting function.
