+++ riroriro/gwexporter.py
@@ -165,6 +165,63 @@ def waveform_exporter_4col(time,freq,Aorth,Adiag,path):
     np.savetxt(path,exportarray,delimiter='\t',newline='\n')
     print('Data saved to %s' % path)
 
+def waveform_arrayer_4col(time,freq,Aorth,Adiag):
+    """
+    Function to collate important data of the simulated gravitational waveform
+    (for SNR calculation etc.) into a single array for ease of storage. Similar
+    to waveform_exporter(), but instead of outputting the area into a file,
+    this function outputs the data inline in Python, in a numpy.ndarray.
+    Includes the two polarisations of strain rather than just the strain
+    envelope amplitude.
+    
+    Parameters
+    ----------
+    time: list of floats
+        The time at each data point. For a BH-BH merger, use i_m_time from
+        time_frequency_stitching in matchingfuns. For a BH-NS or NS-NS merger,
+        use i_time (realtimes) from inspiral_time_conversion in inspiralfuns.
+    freq: list of floats
+        The frequency of the GW signal at each data point. For a BH-BH merger,
+        use i_m_freq from frequency_SI_units in matchingfuns. For a BH-NS or
+        NS-NS merger, use i_freq (freq) from inspiral_phase_freq_integration
+        in inspiralfuns.
+    Aorth: list of floats
+        The orthogonal/plus polarisation of the GW strain at each data point.
+        For a BH-BH merger, use i_m_Aorth from polarisation_stitching in
+        mergersecondfuns. For a BH-NS or NS-NS merger, use i_Aorth (Aorth) from
+        inspiral_strain_polarisations in inspiralfuns.
+    Adiag: list of floats
+        The diagonal/cross polarisation of the GW strain at each data point.
+        For a BH-BH merger, use i_m_Adiag from polarisation_stitching in
+        mergersecondfuns. For a BH-NS or NS-NS merger, use i_Adiag (Adiag) from
+        inspiral_strain_polarisations in inspiralfuns.
+        
+    Returns
+    -------
+    exportarray: numpy.ndarray
+        An array containing the important data of the simulated gravitational
+        waveform: the first column is the time, the second is the frequency,
+        the third is the orthogonal/plus polarisation of strain and the fourth
+        is the diagonal/cross polarisation.
+    """
+    
+    #input type checking
+    for each_variable in locals().values():
+        assert type(each_variable) == list, 'All inputs should be lists.'
+    
+    #checking the lists all have the same length
+    assert len(time) == len(freq) == len(Aorth) == len(Adiag), ('The four '
+                                    'lists all need to have the same length.')
+    
+    exportarray = np.empty((len(time),4))
+    for i in range(len(time)):
+        exportarray[i,0] = time[i]
+        exportarray[i,1] = freq[i]
+        exportarray[i,2] = Aorth[i]
+        exportarray[i,3] = Adiag[i]
+        
+    return exportarray
+
 def waveform_exporter(*args):
     """
     Included to ensure backwards compatibility; redirects to 3col or 4col
@@ -178,3 +235,19 @@ def waveform_exporter(*args):
     else:
         raise TypeError('Inappropriate number of arguments for either the '
             'original or polarisation-included version of the function.')
+
+def waveform_arrayer(*args):
+    """
+    Included to ensure backwards compatibility; redirects to 3col or 4col
+    functions as appropriate.
+    """
+    
+    if len(args) == 3:
+        exportarray = waveform_arrayer_3col(args[0],args[1],args[2])
+    elif len(args) == 4:
+        exportarray = waveform_arrayer_4col(args[0],args[1],args[2],args[3])
+    else:
+        raise TypeError('Inappropriate number of arguments for either the '
+            'original or polarisation-included version of the function.')
+        
+    return exportarray
