### 2021-01-21 1203e1b
Updating citation of forthcoming paper
