+++ paper/paper.bib
@@ -291,8 +291,8 @@ archivePrefix = {arXiv},
 }
 
 @article{massdistribution,
-	author = {{Ghodla}, Sohan and {van Zeist}, Wouter G.~J. and {Eldridge}, J.~J.},
-	title = {Predicting the O3 gravitational wave mass distribution with BPASS},
+	author = {{Ghodla}, Sohan and {van Zeist}, Wouter G.~J. and {Eldridge}, J.~J. and {Stevance}, Heloise F. and {Stanway}, Elizabeth R.},
+	title = {Predicting the LIGO/VIRGO O3a GW transient mass distributions with BPASS},
 	year = "forthcoming"
 }
 
