### 2021-01-28 bbda147
Listing astropy as a dependency
It's only used by _astropy_init (from the astropy package template) and not in the main body of the code, but to be safe it's probably best to include it here.
