+++ setup.cfg
@@ -21,6 +21,7 @@ install_requires =
     matplotlib
     scipy
     pytest
+    astropy
 
 [options.entry_points]
 console_scripts =
