### 2021-01-29 8ea09ef
Amplitude resampling and regular FFT
The Lomb-Scargle functions all seem unable to handle complex values, so now we try a regular FFT function. The amplitude measurements are resampled to be evenly spaced in time because that is required for the FFT function.
