### 2021-01-29 wvanzeist/riroriro#10 @katiebreivik
astropy is missing from pyproject.toml
astropy should be added since it is a required dependency for the package.
Ref: https://github.com/openjournals/joss-reviews/issues/2968
