### 2021-01-30 wvanzeist/riroriro#10 @wvanzeist
Astropy, as well as other non-default dependencies like scipy, pytest etc. is listed in the dependencies list in setup.cfg (lines 19-24), which are automatically checked/installed when riroriro is installed by pip. The TOML file only has default dependencies for the installation itself. Though, astropy was added to setup.cfg after the latest PyPi build, so it won't automatically check for that _yet_.
That should satisfy the condition of it being listed as a required dependency, right?
