### 2021-01-30 2ae4205
Trying to fix Sphinx error with function signatures
