### 2021-02-01 wvanzeist/riroriro#13 @GregoryAshton
JOSS: higher-level comments
Dear Wouter,

I am an assigned reviewer for the JOSS paper associated with this repository (https://github.com/openjournals/joss-reviews/issues/2968).  I'm midway through reading and verifying the functionality and I have a question about the scope. Since this is an open review process, I felt it easier for me to ask this first before addressing the other minor technical details.

If I understand, broadly this `riroriro` provides a user-friendly software package implementing the Buskirk & Babiuc Hamilton (2019) and Huerta et al. (2017) transient gravitational-wave models, with estimates of the SNR passed Barrett et al. (2018) approach. This is useful, but I think the statement of need is not yet clear.

As you briefly describe in the statement of need, there are a number of sophisticated waveform models available in the literature.  These models are well developed and are broadly classified by the "physics" encoded. For example, the inclusion of the merger and ringdown, spin effects, precession, and matter effects have been critical developments in theory with impact on real observations. You may wish to look at modern black-hole waveform models such as the [IMRPhenomXPHM models](https://arxiv.org/pdf/2004.06503.pdf) and [SEOBNRv4PHM](https://arxiv.org/abs/2004.09442) models for a recent review.

As far as I can tell, the model implemented in `riroriro` is not a sophisticated waveform model and contains few (if any) of the sophisticated physics needed to understand observations. This is by design. `riroriro` is predominantly used for predicting the SNR distribution of populations of events for which one needs only an approximate waveform model. However, this is not described in the text. Instead, one gets the impression this attempts to better implement sophisticated waveform models. Additionally, there seems to be confusion between the waveform models and the parameter estimation software (these are separate entities). Please can you consider if your statement of need is accurate and add details about the physics described in the implemented waveform?

Finally, the statements

> The gravitational wave observatories LIGO/Virgo use their own models as templates, and various other research groups have also created models.

and

> However, for the most part these articles do not go into much detail about the computational implementation of their simulations and have not published their full database or its source code. Some have published it but in a language that is not free to use like the original model of Buskirk & Babiuc-Hamilton (2019) which is written in Mathematica.

Are incorrect, ambiguous, and not properly referenced. Phrases such as "for the most part" and "some" are ambiguous: which specifically are you referring to? The models used by the LIGO/Virgo collaborations (not the observatories themselves) are openly accessible (https://git.ligo.org/lscsoft/lalsuite) and use a mixture of C and python.  More troublingly, the models are described in a number of peer-reviewed papers (referenced throughout any detection paper which uses them).

I believe these points need to be addressed before this paper can be accepted. Fairly placing this work in context is critical.

I welcome any discussion on this topic, but please be advised this is a particularly busy period so it may take me some time to respond.

All the best,

Greg
