+++ setup.cfg
@@ -41,7 +41,7 @@ testpaths = "riroriro" "docs"
 astropy_header = true
 doctest_plus = enabled
 text_file_format = rst
-addopts = --doctest-rst
+#addopts = --doctest-rst
 
 [coverage:run]
 omit =
