+++ tox.ini
@@ -4,7 +4,7 @@ envlist =
      python3.7-test-numpy{116,117,118}
 
 [testenv]
-setenv = PYSYN_CDBS = {toxinidir}/hoki/data/cdbs/
+setenv = PYSYN_CDBS = {toxinidir}/riroriro/data/cdbs/
 deps =
     numpy116: numpy==1.16.*
     numpy117: numpy==1.17.*
