### 2021-02-03 a91ddf7
Fixing Tox errors from tests
The mergerfirstfuns test was off because the m_time in the assert statement was for a different "generic system" which I had later changed.
