+++ riroriro/tests/test_horizondistfuns.py
@@ -25,7 +25,7 @@ def test_horizondistfuns_numerical():
     numerically.
     """
     
-    import urllib
+    import urllib.request
     
     #we load in the three files of example data; ideally np.genfromtxt would be
     #used for this but these files are stored on the riroriro_tutorials github
