### 2021-02-03 53aaccc
Re-adding build_docs info
For the Sphynx documentation.
