+++ tox.ini
@@ -12,3 +12,11 @@ deps =
     pytest
 commands =
     pytest
+
+[testenv:build_docs]
+changedir = docs
+description = invoke sphinx-build to build the HTML docs
+extras = docs
+commands =
+    pip freeze
+    sphinx-build -W -b html . _build/html
