### 2021-02-03 wvanzeist/riroriro#10 @katiebreivik
Yes - apologies for mixing up the TOML and setup.cfg! This is good now!
