### 2021-02-04 fa17041
Clarifying JOSS paper text
