+++ paper/paper.bib
@@ -332,3 +332,46 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019PASP..131b4503B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@misc{lalsuite,
+      author         = "{LIGO Scientific Collaboration}",
+      title          = "{LIGO} {A}lgorithm {L}ibrary - {LALS}uite",
+      howpublished   = "free software (GPL)",
+      doi            = "10.7935/GT1W-FZ16",
+      year           = "2018"
+}
+
+@ARTICLE{imrphenom,
+       author = {{Pratten}, Geraint and {Garc{\'i}a-Quir{\'o}s}, Cecilio and {Colleoni}, Marta and {Ramos-Buades}, Antoni and {Estell{\'e}s}, H{\'e}ctor and {Mateu-Lucena}, Maite and {Jaume}, Rafel and {Haney}, Maria and {Keitel}, David and {Thompson}, Jonathan E. and {Husa}, Sascha},
+        title = "{Let's twist again: computationally efficient models for the dominant and sub-dominant harmonic modes of precessing binary black holes}",
+      journal = {arXiv e-prints},
+     keywords = {General Relativity and Quantum Cosmology},
+         year = 2020,
+        month = apr,
+          eid = {arXiv:2004.06503},
+        pages = {arXiv:2004.06503},
+archivePrefix = {arXiv},
+       eprint = {2004.06503},
+ primaryClass = {gr-qc},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv200406503P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{seobnr,
+       author = {{Ossokine}, Serguei and {Buonanno}, Alessandra and {Marsat}, Sylvain and {Cotesta}, Roberto and {Babak}, Stanislav and {Dietrich}, Tim and {Haas}, Roland and {Hinder}, Ian and {Pfeiffer}, Harald P. and {P{\"u}rrer}, Michael and {Woodford}, Charles J. and {Boyle}, Michael and {Kidder}, Lawrence E. and {Scheel}, Mark A. and {Szil{\'a}gyi}, B{\'e}la},
+        title = "{Multipolar effective-one-body waveforms for precessing binary black holes: Construction and validation}",
+      journal = {Phys. Rev. D},
+     keywords = {General Relativity and Quantum Cosmology},
+         year = 2020,
+        month = aug,
+       volume = {102},
+       number = {4},
+          eid = {044055},
+        pages = {044055},
+          doi = {10.1103/PhysRevD.102.044055},
+archivePrefix = {arXiv},
+       eprint = {2004.09442},
+ primaryClass = {gr-qc},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020PhRvD.102d4055O},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
