+++ paper/paper.md
@@ -30,7 +30,7 @@ The gravitational waveform simulation of `Riroriro` is based upon the methods of
 
 Gravitational waves have long been an area of research in astronomy, and particularly since the first observation of gravitational waves was announced in 2016 [@gw150914discovery] this area has seen a lot of research activity. Observations of gravitational waves from binary mergers can provide unique information about their progenitors and stellar populations, especially when combined with electromagnetic observations in the field called multi-messenger astronomy. A major factor in the successful detection and analysis of gravitational wave signals is the creation of simulations of such signals which observed data can be compared to. Because of this, multiple gravitational wave models have been created over the years.
 
-In particular, the gravitational wave observatories LIGO and Virgo have created their own models to use as templates in gravitational wave searches, with the main software for this being `LALSuite` [@lalsuite]. Various other research group have also created models, with some examples of recent sophisticated waveform models being `IMRPhenomXPHM` [@imrphenom] and `SEOBNRv4PHM` [@seobnr]. The waveform model of `Riroriro` that is based on @huerta2017 and @buskirk2019 is not as detailed or precise as these, but it is sufficiently accurate for use in evaluating the SNRs and detectability of populations of events. The code of `Riroriro` is also structured and commented in such a way that users could substitute in functions from other algorithms or even use the detectability modules on output waveforms from other sources, as long as these are formatted in the right way.
+In particular, the gravitational wave observatories LIGO and Virgo have created their own models to use as templates in gravitational wave searches, with the main software for this being `LALSuite` [@lalsuite]. Various other research group have also created models, with some examples of recent sophisticated waveform models being `IMRPhenomXPHM` [@imrphenom] and `SEOBNRv4PHM` [@seobnr]. The waveform model of `Riroriro` that is based on @huerta2017 and @buskirk2019 is not as detailed or precise as these, but it is sufficiently accurate for use in evaluating the SNRs and detectability of populations of events. The code of `Riroriro` is also structured and commented in such a way that users could substitute in functions from other algorithms or even use the detectability modules on waveforms from other sources, as long as these are formatted in the right way.
 
 `Riroriro` combines areas covered by previous models (such as gravitational wave simulation, SNR calculation, horizon distance calculation) into a single package with broader scope and versatility in Python, a programming language that is ubiquitous in astronomy. Aside from being a research tool, `Riroriro` is also designed to be easy to use and modify, and it can also be used as an educational tool for students learning about gravitational waves.
 
