+++ paper/paper.bib
@@ -336,7 +336,7 @@ archivePrefix = {arXiv},
 @misc{lalsuite,
       author         = "{LIGO Scientific Collaboration}",
       title          = "{LIGO} {A}lgorithm {L}ibrary - {LALS}uite",
-      howpublished   = "free software (GPL)",
+      howpublished   = "Free software (GPL)",
       doi            = "10.7935/GT1W-FZ16",
       year           = "2018"
 }
