### 2021-02-05 354ddd6
More paper text improvements
