+++ paper/paper.bib
@@ -343,7 +343,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{imrphenom,
        author = {{Pratten}, Geraint and {Garc{\'i}a-Quir{\'o}s}, Cecilio and {Colleoni}, Marta and {Ramos-Buades}, Antoni and {Estell{\'e}s}, H{\'e}ctor and {Mateu-Lucena}, Maite and {Jaume}, Rafel and {Haney}, Maria and {Keitel}, David and {Thompson}, Jonathan E. and {Husa}, Sascha},
-        title = "{Let's twist again: computationally efficient models for the dominant and sub-dominant harmonic modes of precessing binary black holes}",
+        title = "{Let's twist again: Computationally efficient models for the dominant and sub-dominant harmonic modes of precessing binary black holes}",
       journal = {arXiv e-prints},
      keywords = {General Relativity and Quantum Cosmology},
          year = 2020,
