### 2021-02-09 d62b616
Fixing type error
The FFT output this gives *almost* looks good, definitely better than the Lomb-Scargle ones, but it has some glitchy behaviour at high frequencies in the spectrum which might be related to the resampling.
