### 2021-02-10 64e3333
Doubling FFT resampling frequency
In order to reduce glitches that were occurring at high frequencies in the FFT when the lower frequency was used.
