### 2021-02-10 bdcbfc9
Removing obsolete warning in docstring
The warning is no longer necessary because the reworking mentioned was implemented (in merger_polarisations).
