### 2021-02-12 ae79254
Adapting amplitude_interpolation for new FFT
This (and thereby the subsequent functions also) can now accept input from either findchirp_fourier or proper_fourier.
