### 2021-02-12 3c1474e
Tweaking proper_fourier outputs
Removing unnecessary negative frequencies from the FFT from the output and returning the FFT amplitudes as absolute values instead of complex numbers.

Though, the resultant amplitudes seem nearly a factor of ten higher than the equivalent values from the FINDCHIRP-based FFT, which is an excessively large difference.
