### 2021-02-15 97159be
Further edits to paper text
