### 2021-02-15 03700f1
Minor paper text addition
