+++ paper/paper.md
@@ -41,7 +41,7 @@ Features of `Riroriro` include:
 - Simulating the gravitational waveform signal from a binary merger of two black holes, two neutron stars or a black hole and a neutron star and outputting the data of this signal in terms of frequency and strain amplitude.
 - Using a gravitational wave output and given a detector noise spectrum (such spectra are made publicly available by LIGO), calculating the signal-to-noise ratio (SNR) of the signal at a given distance assuming optimal alignment.
 - Calculating the horizon distance (maximum distance at which an event could be observed) for a gravitational wave model and a given detector.
-- Given the optimal-alignment SNR of an event, evaluating its detectability, the probability that the event would be detected with a SNR above the commonly used threshold of 8, if the alignment would be arbitrary.
+- Given the optimal-alignment SNR of an event, evaluating its detectability, the probability that the event would be detected with a SNR above the commonly used threshold of 8, if the alignment would be arbitrary. These results could then be combined with population synthesis calculations to estimate how many of the predicted mergers would be detected.
 
 In addition, we have created Jupyter Notebook tutorials to help users get started with `Riroriro` ([see tutorials](https://github.com/wvanzeist/riroriro_tutorials)).
 
