### 2021-02-15 ea39b13
Updating PDF and pruning bib file
References not used by the current version of the paper have been removed from the bib file for brevity.
