+++ paper/paper.bib
@@ -80,140 +80,6 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{apostolatos1995,
-       author = {{Apostolatos}, Theocharis A.},
-        title = "{Search templates for gravitational waves from precessing, inspiraling binaries}",
-      journal = {Phys. Rev. D},
-     keywords = {04.80.Nn, 04.30.-w, 97.60.Jd, 97.60.Lf, Gravitational wave detectors and experiments, Gravitational waves: theory, Neutron stars, Black holes},
-         year = 1995,
-        month = jul,
-       volume = {52},
-       number = {2},
-        pages = {605-620},
-          doi = {10.1103/PhysRevD.52.605},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/1995PhRvD..52..605A},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{buonanno2003,
-       author = {{Buonanno}, Alessandra and {Chen}, Yanbei and {Vallisneri}, Michele},
-        title = "{Detection template families for gravitational waves from the final stages of binary black-hole inspirals: Nonspinning case}",
-      journal = {Phys. Rev. D},
-     keywords = {04.30.Db, 04.25.Nx, 04.80.Nn, 95.55.Ym, Wave generation and sources, Post-Newtonian approximation, perturbation theory, related approximations, Gravitational wave detectors and experiments, Gravitational radiation detectors, mass spectrometers, and other instrumentation and techniques, General Relativity and Quantum Cosmology},
-         year = 2003,
-        month = jan,
-       volume = {67},
-       number = {2},
-          eid = {024016},
-        pages = {024016},
-          doi = {10.1103/PhysRevD.67.024016},
-archivePrefix = {arXiv},
-       eprint = {gr-qc/0205122},
- primaryClass = {gr-qc},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2003PhRvD..67b4016B},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{ajith2007,
-       author = {{Ajith}, P. and {Babak}, S. and {Chen}, Y. and {Hewitson}, M. and
-         {Krishnan}, B. and {Whelan}, J.~T. and {Br{\"u}gmann}, B. and
-         {Diener}, P. and {Gonzalez}, J. and {Hannam}, M. and {Husa}, S. and
-         {Koppitz}, M. and {Pollney}, D. and {Rezzolla}, L. and
-         {Santamar{\'i}a}, L. and {Sintes}, A.~M. and {Sperhake}, U. and
-         {Thornburg}, J.},
-        title = "{A phenomenological template family for black-hole coalescence waveforms}",
-      journal = {Classical and Quantum Gravity},
-     keywords = {General Relativity and Quantum Cosmology, Astrophysics},
-         year = 2007,
-        month = oct,
-       volume = {24},
-       number = {19},
-        pages = {S689-S699},
-          doi = {10.1088/0264-9381/24/19/S31},
-archivePrefix = {arXiv},
-       eprint = {0704.3764},
- primaryClass = {gr-qc},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2007CQGra..24S.689A},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{buonanno1999,
-       author = {{Buonanno}, Alessandra and {Damour}, T.},
[…]
