### 2021-02-15 wvanzeist/riroriro#13 @wvanzeist
Thank you for your comments; the text of the JOSS paper has now been clarified and revised in light of these. Could you please take a look at the updated text and see if this addresses your concerns?
