### 2021-02-26 30efdca
Correcting cosmetic error in angle definition
The mistake that was present in the definition of phi did not have any effect on the output because phi is only used in the form cos(2*phi) or sin(2*phi) anyway, but this is the technically correct way of defining phi.
