### 2021-02-26 719a1c3
Function for SNR for specific orientation
