### 2021-03-02 ed98c11
Update version to 0.1.7
Includes the new specific_orientation_SNR function.
