### 2021-03-02 wvanzeist/riroriro#13 @GregoryAshton
@wvanzeist I'm sorry for the delay. Again, I'm afraid this is a busy period with teaching so I'm unable to give this my full attention on the timescale you want.

The new text better sets the software in the context of the field, but to me doesn't yet do this clearly. For example, IMRPhenomXPHM and SEO are both packages available in lalsuite. But, this is somewhat beside the point. The main message is now clear: that the waveform in this package is appropriate for the scientific context. As such, I'm happy to close this issue and recommend the paper for publication,
