### 2021-03-02 wvanzeist/riroriro#18 @GregoryAshton
Remaining JOSS issues
Hi @wvanzeist can you clarify the following points for the final steps of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2968)

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I have ticked the box, because the documentation does exist [here](https://github.com/wvanzeist/riroriro/tree/main/docs/riroriro). But, one has to go and search for it. I'd recommend linking to it from the top-level README to help users. As an aside, you may also want to consider automating the creation and hosting of this documentation using, for example, sphinx and github action.

> Are there automated tests or manual steps described so that the functionality of the software can be verified?

There are no automated tests, but I do find [this directory](https://github.com/wvanzeist/riroriro/tree/main/riroriro/tests). However, the steps to run these tests are not described except briefly [here](https://github.com/wvanzeist/riroriro/blob/main/docs/riroriro/tests.rst). I don't find this enough for a new user to check that they have installed the software correctly. Can you please add something describing how to run the tests?

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I don't see any guidelines for this on the top-level README. Could you either point me to the guidelines, or add them?
