### 2021-03-02 6c136e0
Elaborating on how to run tests with Tox
Re: #18
