### 2021-03-02 f27d47a
Listing sphinx-astropy as dependency
