+++ setup.cfg
@@ -22,6 +22,7 @@ install_requires =
     scipy
     pytest
     astropy
+    sphinx-astropy
 
 [options.entry_points]
 console_scripts =
