+++ setup.cfg
@@ -22,7 +22,6 @@ install_requires =
     scipy
     pytest
     astropy
-    sphinx-astropy
 
 [options.entry_points]
 console_scripts =
