### 2021-03-02 wvanzeist/riroriro#18 @wvanzeist
@GregoryAshton

> I have ticked the box, because the documentation does exist [here](https://github.com/wvanzeist/riroriro/tree/main/docs/riroriro). But, one has to go and search for it. I'd recommend linking to it from the top-level README to help users. As an aside, you may also want to consider automating the creation and hosting of this documentation using, for example, sphinx and github action.

I've put a link to the docs in the README. Thank you for your suggestion, I will try to create a ReadTheDocs site or similar at some point.

> There are no automated tests, but I do find [this directory](https://github.com/wvanzeist/riroriro/tree/main/riroriro/tests). However, the steps to run these tests are not described except briefly [here](https://github.com/wvanzeist/riroriro/blob/main/docs/riroriro/tests.rst). I don't find this enough for a new user to check that they have installed the software correctly. Can you please add something describing how to run the tests?

I've elaborated on this in the documentation you linked, it's quite simple using Tox.

> I don't see any guidelines for this on the top-level README. Could you either point me to the guidelines, or add them?

These are in "contributing.md" (this is also Github's default filename for contributing guidelines).
