+++ README.md
@@ -12,7 +12,7 @@ More information on the pip installation can be found here: https://pypi.org/pro
 
 Tutorials for Riroriro can be found here: https://github.com/wvanzeist/riroriro_tutorials
 
-Full documentation of each of the functions of Riroriro can be found here: https://wvanzeist.github.io/riroriro/
+Full documentation of each of the functions of Riroriro can be found here: https://wvanzeist.github.io/
 
 Riroriro is one of several Python packages associated with **BPASS** (Binary Population And Spectral Synthesis), a suite of programs that simulates the evolution of a population of binary and single-star systems from a wide range of initial conditions. Each of these associated packages are named after native animals of New Zealand. The riroriro (*Gerygone igata*, also known as the grey warbler) is a small bird that can be recognised by its distinctive melodious call but is rarely seen, similarly to how black hole binary mergers are detected by their gravitational wave signals rather than visually.
 
