### 2021-03-07 f9dd9b2
Correcting typo in newest docstring
"descriving" -> "describing"
And giving a bit more detail on the definitions of response function angles.
