### 2021-03-09 786441b
Rearranging sentence for formatting reasons
In order to prevent a formatting bug in the JOSS PDF.
