### 2021-03-11 #2968 @wvanzeist
Great thanks for all the reviewers and editors!

@kyleniemeyer When I go to the theoj.org page for this article, the place where the PDF is supposed to be appears empty, and clicking "download paper" gives "file not found". Might there be a bug of some sort, or is it just taking some time for the website to update?
